server/requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
psycopg2-binary
python-multipart
python-jose[cryptography]
passlib[bcrypt]
bcrypt
torch
transformers
librosa
soundfile
pydantic-settings
python-dotenv
uuid6
g2p-en
